RTS-149 Displayed the wrong name field
Since the name field of terms does not translate we have used the
title module to replace it with a real field that does.  Unfortunately
both are available in the fields list of Views under the same name.
It's really hard to tell which is which.

The replaced name field also has the misfeature of not supporting
direct link to the object, so we have to implement this ourself by
also including the tid field and building the URL manually.