RTS-174 uibx_log()
This introduces the uibx module, a place for us to collect functions
and other stuff we find useful across our modules.

I found prefixing all functions with "uib_api_" or "uib_lib_" to
clumbersome.  My first attempt was to end up with "uib_" as prefix
and simply calling this module "uib".  This did not work.  The
module was simply not recognized by Drupal.  My guess is that's
because we have many other modules with names like uib_foo that
somehow conflicts given how Drupal treat prefixes.

So, I ended up with a "uibx" module instead.