RTS-604 Added lowest level areas, which had been missed, to area list for jur/hf
I had missed the lowest level areas "hf" and "jur" in the list that I set up
with output from "area-walker" (modules/uib_migrate/uib_areas_subset_jurhf.txt).
Because of this, the migrated content did not connect properly, and was
invisible on the site as links to the faculties were missing.
Adding the lowest level areas to the list resolved this.
Note that it should not be necessary to add "hf/en" if "hf" is on the list.