RTS-630 Configure l10n_update to pick up translations from site/all/translations
We're not enabling the l10n_update module yet since that cause failures
during installation and running the update is pretty slow as well.

To update the database with the translations from the site/all/translations
you need to enable the l10n_update module and run 'drush l10n-update'.