RTS-394 Changes to remove superfluous page titles, and use area title
Three titles were previously shown on area pages: site name at the top, node title on the left, and
node title in the node content area.

The superfluous area title of the node content was removed by overriding the title in the panel.
In order to show area title instead of node title when they differ, two variables 'page_title'
and 'page_title_link' were setup in template.php, for use in page.tpl.php. The latter script
was modified to show this instead of site name and skipping the superfluous title on the left when
the titles were the same.

In addition, a variable 'area' which points to the relevant area node was defined, for future use.