RTS-638 Set Norwegian date formats as standard
This saves the suggested date formats in the variables
date_format_{long,medium,short}; and uses the module install
hook to create the corresponding custom formats.

The setup of custom formats using system_date_format_save() isn't
needed for the formats to take effect, but is required for the
formats to display properly on the /admin/config/regional/date-time
page.