RTS-794 Format kicker on news articles
Added some code to template.php to format and add markup to field_uib_kicker

If field_uib_kicker was empty, "news" was entered as default kicker text. node->created was used
as kicker timestamp or node->revision_timestamp if it was larger. The timestamp was formatted
as 'medium' and added with span markup class 'uib_kicker_date' to facilitate styling.