RTS-744 Formatting of byline for news articles
The setting for the field 'field_uib_byline' was changed to get links
to the suthors in the content markup.

Code was added to template.php for news articles picking up author markup
from the field_uib_byline, and formatting this into a new variable
('uib_news_byline') which was added to the content. The original
field_uib_byline was hidden.

The new author line was enclosed in a div class 'uib-news-byline'