RTS-852 Language block made visible in Webdesk and add/article
Previous changes in pages.scss set display:none for the entire header. In
order to make it possible to change language, only the name, site-slogan
and logo were hidden in the same way - not the entire header.

It was pointed out that blind users will still 'see' these items...

Some items were reordered and/or fixed.