RTS-815 Added functionality for creating w2 redirects during migration
Changes were made to uib_migrate in order to enable the uib_redirect module,
and to create redirects duing migration of infopages and navpages.

This was implemented in uib_migrate.inc as a complete() function which is
run at the end of the migration of a xtopic record into a w3 node. Both the
xtopic object and the node object are present at this point and a
redirect, using relevant functions of the uib_redirect module.

Changes were also made to XTopic.inc (add url to object) and uib_migrate.info
(add dependency to uib_redirect)