RTS-957 Fixed preprocess_node refrences to specific fields in article
Previous fixes for article fields which directly referenced and modified
the node render arrays caused errors. This was because of changes during
theming the user interface display of article nodes.

These changes (among them, the grouping of fields) had caused restructuring
of the rendering arrays, and the fixing code in template.php (preprocess_node)
had to be adjusted accordingly.