RTS-148 Teasers for menu children are listed in sidebar second
The commit includes (1) the addition of a small menu block which
includes the first menu level below the current node, (2) setting
up the block to read some node information and displaying one
field along with a node title link.

For areas, the field displayed is 'field_uib_primary_text', for
articles and testimonials, the field displayed is 'field_uib_lead'.