RTS-1038 Top level area menu on article and calendar pages
Since these pages are not area pages, no area menu will be visible
automatically.

The solution checks for current area and obtains the relevant menu
item id (area menu) for the area. This function: 'uib_area__get_curent_mlid()'
is then used when creating the menu block in the 'uib_area_menu_block_blocks()',
thus creating a connection to the area so that the menu will display.