RTS-148 Rebuilt listing of node children teasers
The previous solution used a menu as the basis for picking up information
about node children, and this menu was fixed at a certain level, which
could be a problem.

Also, to remove the overhead of creating a menu, the rebuilt block is based
on RTS-1038 and the menu function 'menu_link_get_preferred()' in
combination with 'menu_navigation_links()' rather than setting up a menu
which was just read and hidden.