RTS-708 Support for tabbed article content in uib_article
h2 headers in uib_articles are recoded using a drupal text filter.
The filter is implemented in a module uib_h2tabs which contains the
main functionality, but there are additional bits of code and
changes that were necessary, as well as css.

The solution scans the text and uses any h2 heades to setup a table
of contents in an ul, and divs for the text. small javascipt onclick
target scripts are used to toggle the visibilty.