RTS-907 Setting up a block for a field_collection (Snarveier): NOT USED
The shortcuts link collection (snarveier) on the employee pages have
been sketched to appear in sidebar second. This was not possible since
the field collection was connected to the node, and hence would appear
together with the node information in the main region.

It was possible to setup a block including the field collection by using
entity functions, but the editing, adding, and deleting functionality was
not possible to achieve within reasonble time.

In order not to abandon the researched solution, it has been put in the
'proto' directory.