RTS-910 Added joint div for group of news archive items
A wrapper div with a class 'uib-newsarchive-section' which contains
the h3 title and the relevant news archive rows were added. The way
the view is currently setup, the submit month and year are used as title.

The addition was implemented by way of a view template:
themes/uib_zen/templates/views-view-unformatted--newsarchive.tpl.php
in which the wrapper class was added, and output if there is a title.