RTS-398 Added wrapper div and class for view "ansatte"
There was no avalibale class for styling a title and its subsequent
rows (capital letter and personal information) in this view.

The div with class="uib-ansatte-section" was added by setting up
themes/uib_zen/templates/views-view-unformatted--ansatte.tpl.php

The class will be present if there is a title to be displayed.