RTS-1131 added handling of path aliases for areas
Since the subpathauto module was not able to handle all subpath aliases,
notably those in the form: jur/persons/faculty (for node/4/persons/faculty)
another solution was necessary.

Added two functions in uib_area.module for setting up the relevant aliases
upon creation of an area node, updating an area node and removing them if
the area node is deleted. Also amended uib_migrate to take this into account.

Removed the subpathauto dependency that was added previously to uib_setup.