RTS-1355 Shortened over-long titles during migration
The file "Name" field may only be 255 chars or an error will be generated.

During migration, the xtopic->title is used for filename and it is
not unfrequently > 255 chars. In some cases it might be suspected
that the title is mistakenly used as description (xtopic->dc_description)

This commit adds two tests for what is to be filename:
1. If it (xtopic->title) > 100 chars, it is copied into description when
   description is empty.
2. If it is > 255 chars, it is cut at end to 255 chars before saving, in
   order to avoid postgre error when inserting into the file_managed table.