RTS-1357 Fixed postgres errors and warnings during migration
The PDOExceptions and postgres errors were due to problems during
import of certain w2 autogenerated navigation pages that did not
have any site information attached. Since navigation pages are
treated as articles in w3, and articles should be connected to areas
this resulted in attempts to save an article connected to an empty
area. Postgres did no accept this.

The fix was to add a prepareRow function in the NavPageMigration class.
In this, a check for empty site info is performed, and those navpages
are rejected.

In addition, some postgres database warnings caused by empty menu
link weights were fixed (setting the weights to 0 in such cases).
In function complete() of the NavPageMigration object.