RTS-1667 Added courses views to create alias function
These views (courses, study-programmes etc) are setup to display
as /node/%/courses etc. Because of this, the page will not exist
when an alias url is used, such as /jur/courses.

These views were added to the list of aliases to create upon node
creation (and deletion) in the uib_area module, the functions:
uib_area__create_area_aliases and uib_area__remove_area_aliases