RTS-1604 Do not show Faculty Department Kids block if empty
This block is made by a view and the block is manually inserted into
the node as rendered html using the views_embed_view() function in
uib_zen_preprocess_node().

I added a test on the rendered html to see if there is a class
'view-content' present, which will only be there when the block
is non-empty.