RTS-1761 Fixed that unpublished w2 content are not migrated
Content of type NavPage was imported regardless of it it was
published in w2 or not. This was caused by a change (RTS-1357) which
did not properly take inheritance into account. Also, it caused
these articles to hav empty article types.

The fix adds a proper call to parent PrepareRow method, which was
lacking.