RTS-#1857 Temporary security update: disabling administrator role
After a reinstall, the user roles are numbered differently from
what is the case on the production server (which can not be reinstalled).

Updating administrator role uses role id rather than role name which
causes the wrong role to be set as administrator when added to
production. For this reason, the administrator role is (temporarily)
disabled.