RTS-#1927 Added check to User migration for already existing users
The synchronization script with Sebra (uib-sebra-users) may add users
and this will later cause problems when migrating users into Drupal.
It causes a PDO exception dues to duplicate user names, a field which
should be unique.

In addition, the users added by the Sebra synch will not have the w2 id
set.

This change adds a check for existing users to uib_migrate and when users
do already exist, it will update/set the w2 id.