RTS-#2047 Fixed file migration, especially in relation to translation of description
Rewrote file migration
The main problem was that the description field was set to translatable, but the texts were
set to language undefined [und], which caused the description not to to be rendered properly.
The language needed to be set to [en] or [nb].
There were also other migration things fixed in that function.

The function was rewritten to:
* take the language into account and set this properly for the description field
* Initialize entity translation for image, audio, and video file entities
* make it possible to update file meta data in migrate or migrate --update
* figure out w2 files of types audio and video and set proper file types
* fix any previously migrated [und] remaining