RTS-#1927 Releveant handling of exceptions added
On production, the script did not complete because a lookup for a w2 id in one
user stub (had been added by migrate from w2) failed. This caused an exception
in uib_migrate and these scrips failed to catch the exception, which
effectively terminated the script. I added a simple exception handler.