RTS-#1184 Hiding tabs for Level 2 and 3 users
The edit area form tabs at the bottom has two tabs which are intended
only for user levels above 3. Previous commits in #994 and #1184
only hid relevant fields for these users, but the tabs remained.

By setting access to false insted of hiding the fields, the tabs
specific to level 1 and level 2 users do not show for lower user
roles.