RTS-#2303 Removed previous one-time fix functionality from xtopic_file
The xtopic_file code had become too complex when I added the functionality
intended to fix the description and file title fields which had been
previously migrated [language set to und]. They did not show up when
viewing the file.

The code was cleaned up, and the one-time fixup functionality was
moved to uib_migrate_update_7000().

The xtopic_file retains the proper language setup for new files, but
existing files will not be touched -- unless specifying the drush
options --force (along with --update if migrating).