RTS-#2303 uib_migrate database fix update script
Previously migrated images had problems when the attached text
(field_uib_description) was viewed. This was because translation was
turned on and language had been set to und for the field rather than
being set to the language of the node. Also, entity translation was
not properly initiated, which made the text not show up when the user
tried to translate it. Nor when viewing the file.