RTS-#2381 Added drush command to fix missing OU in area nodes
Area nodes that lack Organisational Unit "OU" indications may cause
warnings etc.

This commit adds a drush command 'uib-area-connect-ou',
that attempts to connect these area nodes with an OU.
Nodes are updated by using node_save() which is not very fast.

Part of the addition is a modified function uib_area__get_ou_nid()
which will - in the case when the area does not have an OU - pass
through the area parents until it (hopefully) finds an OU.

If an area node (that lacks OU connection) has more than one area
parent, we choose the first OU encountered. We want a warning log
for cases when we make an arbitrary choice.