RTS-2068 Transform menu.module to override the 30 we don't like
The menu module truncated the menu titles to 30 characters (or less
when the "--" prefixes were added). The long area names in w3 were
truncated so that they were not possible to distinguish.

This was hard-coded in a helper function in the menu module, and
the override was done by copying the code from four functions into
this module, and changing their names.

The previous commit c8fb2ad9c00b contains the unmodified code.

The title is still truncated, and set to:
round(original_value * pi)