RTS-#2454 Redirects for all article types during migration
News and Events did not get redirects previously. For this purpose,
the complete function was moved from Class InfoPageMigration to its
parent.

The redirect source string may be in two w2 source fields: url or path,
and in some cases both are non empty. In the latter case, two redirects
are set up. (Previously only url was used.)