RTS-#2424 Removed superfluous media queries 2.0
Remains four unique media queries. One for all deviced with a screen
width at least 480 pixels, one for mobile devices, one for tablets,
and one for desktop devices.

35 out of 39 media queries were removed, as well as numerous
settings of the zen-column-count variable.

As the media queries for mobile and tablet overlapped on 720, which
caused weird behaviour at that width, the mobile query was set to
max width 719.