RTS-#2505 Filter for removal of w3.uib.no in pasted links. Version 2...
Urls pasted by users directly from w3.uib.no needed to be fixed
since that domain is not visible on the web.
The commit adds a small module contaning a filter that removes instances of
"w3.uib.no" from pasted inline links in text.

The commit also includes an uib_setup update function which enables
the module, and a dependency-line in uib_setup.info.