RTS-#2529 Convert certain links in link fields to relative links
Created a field formatter for link fields to be able to alter
the links prior to rendering. The formatter
(hook_field_formatter_info) was placed with the view-hook
(hook_field_formatter_view) in the uib_urlfilter in order to
keep all the url filtering functionality together. In this way,
the text filter in RTS-#2505 and this link formatter are
sharing code.