RTS-#2524 Adds copyright to area slideshow images
The view area_slideshow was modified in order to be able to display
copyright information for the image displayed.

Added a relationship to the image via main media file id, and the
copyright field. The info was added to the rewritten title field
in a separate div with a "copyright" class. Styling remains.