RTS-#914 Populate field_uib_area on article creation
In order for articles to be displayed on all parental areas of an
article, the field_uib_area need to contain all relevant
area parents. In this way, existing views will continue to work
without modifications.

To populate this field upon article creation, an implementation of
hook_node_presave was added to modules/uib_article.module. This
makes use of a function in uib_area.module (uib_area__get_parentage)
that fetches all parental areas of an area.