Add 'site-upgrade' script
It takes a git commmit as argument and will then upgrade the
application to this commit.  This includes checkout, recompilation
of sass files, feature reverts and clearing the caches.

Normally you should only try upgrades from the production code
and database.  Run 'bin/site-prod-reset' to get there.

The 'site-clean' script is at the same time retired.  There should
not be any use for it.  Just 'site-upgrade' where you want to go.