RTS-#914 Adds an opt-out checkbox for area populating in new articles
In order to let authors decline automatic setting of parent areas for
an article they are creating, an opt-out checkbox was added to the
article creation form. It is only shown when editing a new article, and
was put just below the field_uib_area form field.

It was implemented in the existing uib_article implementation of
hook_form alter().  A few corresponding changes were made to the
hook_node_presave() implementation in order to stop the automatic
area population if the checkbox was checked.