RTS-#914 Update function for adding parents to existing articles
As news and event "bubbling" comes in effect for newly created
articles, one might want to populate the parental trail for
existing articles.

This update function deals with news and event articles and more
specifically those events that have not passed, and those news
articles created 30 days ago or less.