RTS-#2587 Breaking apart the big Area menu into one menu per area
Adds a drush command [uib-area-split-menu] that creates a separate
menu for each area, and moves existing menu items into the new
menu.

Note that the top item for each area, i.e. the one that has the
name of the area is not moved, but remains in the original area
menu.

The machine name for the separate menu is "menu-area-<nid>", its
menu name is "area_<url alias>" with "_en" added for English
language areas. The menu description is "Menu for <node title>".

In order to get the second level in the menu to drop down automati-
cally we need to set first level to "expanded". It turns out that
we also need to insert a title attribute option for it to work.