RTS-#2587 Basic changes to get the separate area menus to show
Added a helper function uib_area__get_current_menu() that fetches
the relevant menu (machine name) for an area. There were several
additional changes to accomodate this.

Removed the "top level area menu" menu_block block which should
be replaced by each separate area menu.

Added rendering of the new separate area menu in page.tpl.php. It
(as a block) is rendered in the navigation div.

Made a very slight modification to navigation.scss because the
styling of the menu was narrowly specific to the top level area
menu.

The top level menu links should not link to any other page.
Previously, there was a function in template.php that fixed this
for the big area menu.