RTS-#2587 Creating (and removing) an area menu when creating areas
Modified the already existing hook_node_insert() and
hook_node_delete() to create and delete an area menu.

The area content type was modified to not have a default menu.

Do not add/remove menus to/from content types.
As we choose to handle area menu attachment for articles on our own,
new menus should not be connected to the article or testimonial
content types. Nor should they be removed.