RTS-#3031 Remove display of byline from event articles
A minor bug in RTS-2429 caused the byline field to show up for all
articles if the byline field was filled.

Inspection of the code revealed other problems due to the elseif
introduced by #2429 which would cause the byline to show for events.
Also, some previous strucural change of the content type had moved
the byline field into a field group which made the template code
which was hiding the byline field actually not work. This commit
aims to fix these three problems.