RTS-#2980 Removal of template code showing author as default byline
The introduction of the external author field in #2292 might cause
an unwanted situation where the UiB owner (uid) of the article
would be presented in the byline together with an external author.

To increase flexibility, the code presenting UiB user as byline if
the byline field was empty, is removed.

(This commit is more or less a duplicate of a previous commit, which
had become difficult to apply. After a rebase with master, the
change is now re-commited.)