RTS-1361 Views patch
The problem with Postgres and the media browser turned out to be a
more general problem in views, and how it handles a certain class of
queries.

The solution to the problem had been found http://drupal.org/node/1331056
and a working patch was available. However, this has not yet been
commited to views development code.

SMB rerolled the patch against the current view development branch and
we have applied it to our stable views installation. Before the patch
is added to views, we cannot update views without re-applying the patch.