RTS-3001 Creates front page areas
This commit includes code to create to areas of type frontpage, one
english and one norwegian. The variables holding information of
multilingual frontpages are updated and set to the node values of
the created nodes.

The commit removes some values from features in uib_setup that otherwise
always will create a feature diff in production.