RTS-#3333 Removing extra gray colour in expanded menu
On MSIE 9, the expanded menu showed a gray bg colour on the list
items, which differed from the expected bg colour - one too many.

The cause was a browser specific gradient filter which did nothing
except setting the LI bg colour to an unwanted colour.

Since the filter did nothing else, the issue was solved by removing
the filter.