RTS-#3240 Adding squares to Frontpage view blocks
Long view and subview names forced views to use hash deltas for
the blocks output from "Frontpage profiled articles".

In order to get the title squares, we needed to specify the
corresponding block html ids.

Also made a minor simplification of the scss, thus avoiding to add
the block ids to the scss.