RTS-#3240 Adding heading squares to feeds
The feeds (jobbnorge and uib i media) were formatted in a way that
caused the block title (block subject, really) to be empty, and
instead the title was formatted directly in html. This made the
preprocessing code in uib_zen_preprocess_block skip the block, and
no blue/orange square was set.

In order to get a proper block subject/title, the data input to the
feed formatter was modified, and the block subject set properly.