RTS-3369 Styling update of 'Follow us' links for area pages
The styling of 'Follow us' links in the area pages where unstable and
not complete for smaller screens. This update tries to simplify the styling
and finish the styling for small screen devices.

The styling introduces a breakpoint where there is no longer room for both
title and icon of the five supported social medias. A width narrower than
580 px will only display the links as icons. A width wider will display
the links as title and icon.