RTS-#3240 Squares on Aktuellt field collection in sidebar
Th field collection "field_uib_link_section" is used
for "aktuelle lenker" as well as "kalender - det skjer" on the
aktuelt areas (english and norwegian).

The field collection instance titles are written by the user, and
these titles should be displayed in the second sidebar similarly to
those of the blocks shown there, i.e. with blue or orange squares.

This solution is based on the supposition that the second field
collection instance is the calendar link collection - which should
have orange square.

A complicating factor is that area pages may use the same field in the
footer. Because of this, the sass code had to be modified to restrict
the blue-block and orange-block to the sidebars.