RTS-#3328 No mobile menu button when no area menu
The menu button for mobile devices was always output, regardless
of if there was a current area menu or not. In addition, the styling
of the contaning element "#menu-search-mobile-wrapper" depended on
the button being present, and would partly collaps when the button
was removed.

To remove the button, a check for presence of current area menu was
added to the page template. A height was added to the contaning
element in the sass file for responsive design.