RTS-#3359 Double bottom borders on blocks in sidebar
Some blocks/field collection items residing in sidebar second
lacked the expected double bottom border. These "blocks" are
visually similar, but not the same kind of drupal items.
Because of this, several pieces of sass code need to be modified
in order to get the bottom border in place.

Previous styling (such as RTS-3245) had to be removed/modified
in order to get similar styling in all places. This commit is
based on changes made in RTS-3240 and therefore commited on that branch.

Basically, the styling is applied to blue- or orange-block items
when in region-sidebar-second. However, in the case of the field
collection "field-uib-link-section" which is used on the aktuelt
page, a separate styling had to be applied due to the different
markup.