RTS-#3437 Get rid of unnecessary PHP warnings
In several places in the uib_area module or preprocess_node of the
template.php file, code attempted to access non-existing variables,
objects or array elements. This caused php warnings.

This commit modifies the access and/or adds checks that certain
variables do exist before trying to access.

One message (line 2211 of w3.uib.no/drupal/modules/node/node.module)
was not possible to reproduce locally, and it may not be fixed.