RTS-#3362 Displaying a default image in newspage_recent_news block
The newspage_recent_news block lists news article teasers. When
no main media image is available, an empty space results. Instead,
one wishes to show a default image.

The block is created manually in uib_area.module. This commit checks
if a teaser lacks image and adds a default. This image needs to be
inside a class "field-name-field-uib-main-media-recent-news" in order
to display correctly. The image used is the same that were used
previously in the view recent_news/newsarchive as "no results" image.