RTS-#3167 Adding a title to employee list (view)
The employee list is the output from a view (two subviews). The list
is added to the page by using the views_embed_view() function in
preprocess_node (template.php). It seems that the view title is
ignored when outputting the view in this way, and a title has to be
added manually.

The title "Staff" was introduced in a preprocess_views_view()
function which was added to template.php.