RTS-#3510 Visibility and format of feed block titles
Due to the way the feeds "jobbnorge" and "feed" are formatted, it is
easy to get a block which seems to contain something although there is
no data. Hence, only the title is displayed. A previous change in #3240
caused the addition of title to make the block always show.

In order for the block to get a properly formatted title and still
not show up when empty, the addition of title made in #3240 is undone
here and the title is mysteriously added as before during feed
formatting in uib_area__feed_format_pre_render.

Since the title when added in this way lacks some markup, this
commit adds a post render function that adds markup necessary to get
the blue square preceeding the block title.