RTS-#3567 Allow front page area to have menu
Areas may or may not have menus and the front page area had not been
setup to allow this. Because of this, its menu was not converted to
the new area menu format.

This commit adds "frontpage" to the list of area types to allow
menu, and an update task so it is converted.