RTS-3572 Fix for missing mobile menu
After the new menu design was introduced the mobile menu was missing. The
problem was the function used to check if the area was supposed to get a
menu or not did no longer return any other value than NULL.

By using another function to verify this, the menu reappears.