RTS-#2587 Minor split menu bugfix
During split, it was possible that an external link was added
more than once if the uib-menu-to-relation script was run more
than once.

This commit should fix this problem. Also, some additional changes
were made to increase information in verbose messages (notices).

Finally, a superfluous update task (uib_area) was removed. It does
not need to run as long as the "frontpage" area is allowed to
have a menu (RTS-3567).