RTS-#3585 Permission managing menu.
Adds a new permission setting 'administer uib area menus', and
handling of this in combination with checks if the user is
content manager of the area - before allowing management of an
area menu.

Previously, all of level 1-3 users needed the powerful 'administer
menus' permission set, which was a problem. This is not necessary
after this change.

In addition, the admin/structure menu item in the admin menu is hidden
from level 3 users. They can't do anything there anyway.