RTS-#3562 Avoid creating menus during area migration
Migration of area menu expects no separate area menu to be present.
However, when areas are created, a menu is created in hook_node_insert.
This cause menu migration to fail.

This commit adds a check for w2-id when creating an area node. If
there is one, we are migrating the node, and refrain from creating
a menu. This is later created during menu split.