RTS-3575 Expanding the uib_breadcrumb module
After the new menu architecure was implemented the default breadcrumb
did not display more than two levels of the active trail and the current
area was missing from the active trail.

This patch adds the current area to the active path and follows the
relation put on articles until an item in the active trail is found.