RTS-#3586 Do not allow menu children below depth 2
Area menus are now limited to two levels (depth 1-2). Because of this,
users should not be allowed to add menu items below depth 2.
The uib_article_menu_form in the menu local task allowed this.

This commit adds ajax functionality to the form, which adjust the
radio buttons for menu link relationship accordingly.