RTS-#3648 Adds uib external to features and sets permissions
The permission settings for uib external had not been exported
as feature components, nor had any permissions been set.

This adds permissions for this content type to uib_setup. All of
the permissions are set to allow level 1-3 and admin the same
as for uib article, i.e. all permissions.