RTS-#3705 Translating node form failed with PDO exception
When adding a new node translation, two unrelated things failed.
Both of them caused a PDO exception in itself.

The problem in uib_article_form_alter was due to the incoming node
structure being different from the usual node: it lacked node id.
This commit adds consideration of the alternate node structure.

The problem with the uib_breadcrumb_menu_breadcrumb_alter was that
it failed on pages lacking a current area.
This commit adds a check for if there were a current area.