RTS-#3707 Apparent duplicates of related content
The script that moves menu links at depth > 2 into a structure of
related content pointers had an issue when handling hidden menu
links: If the hidden menu link had children below it that were not
hidden in themselves, the child links would be used.

This commit changes the function in a way that all links from a hidden
menu link and all below it will be ignored, and removed.

It also fixes a potential problem where an external content
node would be created even if the external link had been hidden.