RTS-#3277 Adding markup to testimonial teaser quote and name
Since the styling of the testimonial teaser does not follow the grid,
and is very sensitive to any kind of change, it was considered impractical
to add quotes by using css. The changes were made by adding to the
markup in template.php instead, in combination with modifications of the
content type display for the teaser view mode.

Two of the fields had been setup as having field labels in the view mode
and hidden by using css. Instead I removed the lables from the display and
also removed the unnecessary css that hid the labels.