RTS-#3729 Changes to make sync_w2_users play well with xtopic_file
The functionality of, and interface to, the xtopic_file function
have changed without corresponding updates to the w2 user sync
function.

This commit aims to accomodate this.

Note that the user image is updated from w2 if:
1. There is no previous image in w3.
2. The w2 id of the image in w2 is different from the one in w3.

It is said to be possible to change user image in w2 without change
of w2 id.