RTS-#3763 Do not try to set redirect if longer than 255 characters
Redirect sources longer than 255 characters cause a PDSO exception,
for instance when migrating from w2.

It would not work to truncate the redirect source if it is too long,
since the idea is to get a redirect which works with the old w2 path.

Thus, this commit checks source length, skips the redirect entirely,
and outputs a warning message.