RTS-3785 Make labels consistent and more styleable in article sidebar
The labels used above fields in the article sidebar where not displayed
the same way. This made it difficult to style and caused labels not
intended to to be displayed in the same font as labels above fields.

This commit makes labels that are supposed to be displayed above the
field to get the correct classes and rewrite the style rules setting
the font family to not include all labels in article sidebar.