RTS-3840 Make active_path more aware of context.
The breadcrumb module searched for related content and picked the first
it found to build the breadcrumb. This could be the wrong relation and
displays some odd results when the article is placed in several locations.

This fix searches for related content (and items in the menu) that belongs
to the same area as the article is displayed.