RTS-#3754 Adjusting area page sass to follow zen grid columns.
Changes for Desktop screens (width > 959px).

Most non-layout css are moved to other files. I have not tampered with
(much of) the css for the slide show or the primary media fix (that
incleases the size of the media when no primary text is available).

Note that the blocks now show in the order they are rendered. I have
not changed the order since it would affect the order on other
devices.

Note also that when placed in the grid, adjacent blocks touch. If
they have a background colour, there is no space in between. This
is because the standard zen grids gutter is implemented as padding.