RTS-#3754 Adjusting area page for tablet and mobile to fit zen grid
Displayed fields, blocks and views were fit on the zen grid, and
in most cases, zen grid item base indents were applied.

The group_two_column area was attached to the grid as a whole,
and all fields/views internal to it were set to simply float
left within it. This was done in order to have a better handling of
cases where one or more internal fields were missing.

The following were not adjusted due to their complex and
easy-to-break formatting:
- primary media size change when no primary text is present (includes
  manual handling in template preprocessing)
- slide show

Some of the changes for tablet and mobile made further changes for
desktop also desirable. Thus, some of the changes in this commit
are related to desktop-sized area page display.