RTS-#1069 Added date title override for Day calendar
The Day view (a uib_calendar view) displays calendar items for a single
day below a title with a default format date. This is setup by
theme_date_nav_title theming function in the date module.

We override that theming function in order to set our preferred
format on the title.