RTS-3553 Adding a new calendar displaying year
Adding 2 new displays, event filter year and year. Tried to use the old event filter, but this filter only
filtered on year-month, so needed a new one for year.
The link to year is displayed on the mini calendar, this uses a custom url. Tried to use the year display link, but this passed the month as variable
The php filter code was entered manually for both event filter and event filter year, something still wrong with features.