RTS-#3930 Styling of fields etc in group_two_column
The commit includes two things:
1. The embedded faculty_department_kids view needed to be in the
new group_column_first.

2. Styling of the contents of the two new groups group_column_first
and group_column_second, placing them in the zen grid. The styling
for desktop and tablet are idential, while the styling for mobile
differs.

Unfortunately, due to the way the media queries are setup
in sass/layout/responsive-sidebars.scss the identical code for
styling on desktop and tablet needs to be in two places.