RTS-3972 Make breadcrumb show context of the menu first
When a node exists both as a menu item and as related content in an area
both top level menu items were displayed. This fix makes sure that if a
node exists both as menu item and as related content, the path for the
menu item is displayed as the breadcrumb.